memstore: don't encode/decode 'size_t'
there is no encode/decode funtions for type 'size_t'. We get following
error on OSX:

os/PageSet.h:207:5: error: call to 'encode' is ambiguous
os/MemStore.h:144:7: error: call to 'encode' is ambiguous

+++ src/os/PageSet.h
@@ -98,7 +98,7 @@ class PageSet {
   typedef typename page_set::iterator iterator;
 
   page_set pages;
-  size_t page_size;
+  uint64_t page_size;
 
   typedef Spinlock lock_type;
   lock_type mutex;
